feat: migrate ProfileDashboard to Dashboard component and restructure dashboard-related hooks and components
