+++ src/pages/dashboard/components/ProfitWidget.tsx
@@ -1,17 +1,17 @@
 import clsx from 'clsx';
-import { UnitSales } from '../../../../models/statistics.interface';
-import { DashboardUnit, normalizeSoldGamesData } from '../../../../utils/statistics';
+import { UnitSales } from '../../../models/statistics.interface';
+import { DashboardUnit, normalizeProfitData } from '../../../utils/statistics';
 import { Equal, TrendingDown, TrendingUp } from 'lucide-react';
+import { formatPrice } from '../../../utils/formatPrice';
 
-export function SoldGamesWidget({ data, timeUnit }: { data: UnitSales; timeUnit: DashboardUnit }) {
-  const { soldGames, difference, percent, delta } = normalizeSoldGamesData(data);
+export function ProfitWidget({ data, timeUnit }: { data: UnitSales; timeUnit: DashboardUnit }) {
+  const { profit, percent, delta } = normalizeProfitData(data);
 
   return (
     <article className="bg-placeholder space-y-4 rounded p-6">
-      <h4>Juegos Vendidos</h4>
-
+      <h4>Total Ganado</h4>
       <div className="font-bebas flex gap-4 text-xl">
-        <span className="text-4xl">{soldGames}</span>
+        <span className="text-4xl">{formatPrice(profit, false)}</span>
         <span
           className={clsx({
             'text-green': delta === 'increase',
@@ -26,11 +26,11 @@ export function SoldGamesWidget({ data, timeUnit }: { data: UnitSales; timeUnit:
 
       {delta !== 'same' && (
         <p>
-          Se {difference > 1 ? 'vendieron' : 'vendió'} <strong>{difference}</strong>{' '}
-          {difference > 1 ? 'juegos' : 'juego'} {delta === 'increase' ? 'más' : 'menos'} que el {timeUnit} pasado
+          Las Ganancias {delta === 'increase' ? 'aumentaron' : 'disminuyeron'} en un <strong>{percent}%</strong>{' '}
+          respecto al {timeUnit} pasado
         </p>
       )}
-      {delta === 'same' && <p>Se vendió la misma cantidad de juegos que el {timeUnit} pasado</p>}
+      {delta === 'same' && <p>Las ganancias fueron las mismas que el {timeUnit} pasado</p>}
     </article>
   );
 }
